pyproject.toml
[tool.poetry]
name = "brainsmith"
version = "0.1.0"
description = "Backend project for brainsmith"
authors = ["Daniel Dai <dch950429@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.11"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.0"
langchain-text-splitters = "^0.3.4"
python-multipart = "^0.0.20"
langchain-ollama = "^0.2.2"
langchain-chroma = "^0.1.4"
langchain-openai = "^0.2.14"
streamlit = "^1.41.1"
langgraph = "^0.2.60"
langchain-experimental = "^0.3.4"
pypdf = "^5.1.0"
redis = "^5.2.1"
docx2txt = "^0.8"
bs4 = "^0.0.2"
validators = "^0.34.0"
itsdangerous = "^2.2.0"
python-jose = "^3.3.0"
authlib = "^1.4.1"
pyjwt = "^2.10.1"
sqlalchemy = "^2.0.37"
passlib = "^1.7.4"
email-validator = "^2.2.0"
psycopg2-binary = "^2.9.10"


[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ipython = "^8.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
